#151: move ballotPaperSectionsPathObject in correct folder
